// src/app/modules/main-container/models/message-record.ts
import {Visitor} from "./visitor";
import {Reply} from "./reply";

export class MessageRecord {
  id: number;
  contentId: number;
  title: string;
  content: string;
  regDate: string;
  regTime: number;
  viewCount: number;
  messageTypeId: number;
  regUser: string;
  profileAddress: string;
  fullname: string;
  fullName: string;
  parentId: number;
  parentContentId: number;
  album: string;
  attachments: string;
  isSelf: boolean;
  isCreator: boolean;
  isAdmin: boolean;
  isVisit: boolean;
  isDelivered: boolean;
  creator: string;
  orgContent: string;
  receiverId: string;
  uniqueId: string;
  sourceId: number;
  isObsolete: boolean;
  element: any;
  isTotalView: boolean;
  isPartialView: boolean;
  isLongTotalView: boolean;
  height: number;
  itemId: string;
  innerHtml: string;
  dayName: string;
  isFirstMessageOfDay: boolean;
  isSelected: boolean;
  isHover: boolean;
  messageId: number
  visitors: Visitor[] = [];
  replyMessage: Reply = new Reply();
  isFirst: boolean = false;
  constructor() {
    this.id = -1;
    this.contentId = -1;
    this.title = '';
    this.content = '';
    this.regDate = null;
    this.regTime = 0;
    this.viewCount = 0;
    this.messageTypeId = 0;
    this.regUser = '-';
    this.profileAddress = '-';
    this.fullname = '-';
    this.parentId = 0;
    this.parentContentId = 0;
    this.album = '-';
    this.attachments = '-';
    this.isSelf = false;
    this.isCreator = false;
    this.isAdmin = false;
    this.isVisit = false;
    this.isDelivered = false;
    this.creator = '-';
    this.orgContent = '-';
    this.receiverId = '-';
    this.uniqueId = '-';
    this.sourceId = 0;
    this.isObsolete = false;
    this.element = null;
    this.isTotalView = false;
    this.isPartialView = true;
    this.isLongTotalView = false;
    this.height = 0;
    this.itemId = '';
    this.innerHtml = '';
    this.dayName = '';
    this.isFirstMessageOfDay = false;
    this.isSelected = false;
    this.isHover = false;
    this.messageId = -1;
    this.isFirst=false;
  }
}